discovery: extract MOLS policy into selectors/mols package
- Add portal/discovery/selectors/mols with type MOLS (renamed from
  MOLSRelayPolicy), constructor New(), and compile-time assertion
  var _ discovery.Selector = (*MOLS)(nil).

- Define local unexported relayPolicy interface in portal/discovery to
  break the import cycle; RelaySet.policy is now relayPolicy, and
  NewRelaySet accepts a relayPolicy parameter (callers pass mols.New()).
  isNilableAndNil helper guards against typed-nil injection.

- Rename RelayState.hasObservedDescriptor() -> HasObservedDescriptor()
  and add IsSuppressedActive(now time.Time) bool for cross-package access.

- Migrate all call sites (portal/server.go, sdk/expose.go,
  sdk/expose_test.go, cmd/portal-loadtest/main.go) to mols.New().

- Add package-discovery stubRelayPolicy in testhelpers_test.go for
  white-box tests that cannot import selectors/mols (cycle).

- Move math and policy tests to selectors/mols/mols_test.go (package mols,
  white-box); keep suppress/backoff tests in portal/discovery/mols_test.go
  (require unexported RelayState fields); keep lifecycle tests in
  portal/discovery/policy_test.go.

- Delete portal/discovery/mols.go. Trim portal/discovery/mols_test.go to
  the two tests that access unexported RelayState fields. Remove policy
  tests from policy_test.go that are now covered in selectors/mols.

go build ./... && go vet ./... && go test ./... all pass (149 tests).
git grep MOLSRelayPolicy -- ':!docs' ':!*.md' returns zero matches.
